feat: seperate game form to components
