+++ frontend/components/admin_form/game/edit_form.tsx
@@ -1,0 +1,7 @@
+export default function GameEditForm() {
+    return (
+        <form className="flex flex-col gap-6 font-title">
+            test
+        </form>
+    )
+}
